+++ NEWS.md
@@ -1,3 +1,15 @@
+# bestNormalize 1.6.7
+
+- changed `step_bestNormalize` to `step_best_normalize`, responding to [8](https://github.com/petersonR/bestNormalize/issues/8)
+- Fixed error in documentation regarding `LambertW` transformation types 
+  (thank you to Georg M. Goerg, the author of `LambertW`, for pointing this out). 
+- Add `center_scale` transform as default when `standardize == TRUE`
+- Added error when trying to use repeated CV with much too small of folds 
+- Changed a few `T` and `F` to `TRUE` and `FALSE` 
+- Added documentation of how one can use `scales` and `ggplot2` to visualize all transformations.
+- Added `butcher` and `axe` functionality in order to improve scalability of `step_*` functions
+- Improved `tidy` functionality with bestNormalize and `step_best_normalize`
+
 # bestNormalize 1.6.1
 
 - Fixed bug that was causing simple transforms to fail in `bestNormalize`